Rename DeleteCommandParser to DeleteContactCommandParser and repackage it
