+++ src/main/java/seedu/address/logic/parser/contacts/DeleteContactCommandParser.java
@@ -1,15 +1,17 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.contacts;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.contacts.DeleteContactCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
  * Parses input arguments and creates a new DeleteContactCommand object
  */
-public class DeleteCommandParser implements Parser<DeleteContactCommand> {
+public class DeleteContactCommandParser implements Parser<DeleteContactCommand> {
 
     /**
      * Parses the given {@code String} of arguments in the context of the DeleteContactCommand
